fix: to avoid web container Errors when using Elastic-APM
uwsgi.ini:
The change to uwsgi's enable-threads was done according to the
instructions in the Elastic-APM document.

Dockerfile:
When I built it in my environment, 3.13 was downloaded, which caused a
gzip error, so I downgraded to 3.12. According to the python document,
both are stable versions.

gzip error detail:
Exception ignored in: <gzip on 0xffffa8ffd600>
Traceback (most recent call last):
  File "/usr/local/lib/python3.13/gzip.py", line 359, in close
    fileobj.write(self.compress.flush())
ValueError: I/O operation on closed file.